+++ core/src/main/scala/org/bitcoins/core/crypto/TransactionSignatureSerializer.scala
@@ -1,7 +1,7 @@
 package org.bitcoins.core.crypto
 
 import org.bitcoins.core.currency.CurrencyUnit
-import org.bitcoins.core.number.UInt32
+import org.bitcoins.core.number.{Int32, UInt32}
 import org.bitcoins.core.protocol.CompactSizeUInt
 import org.bitcoins.core.protocol.script._
 import org.bitcoins.core.protocol.transaction._
@@ -114,7 +114,7 @@ sealed abstract class TransactionSignatureSerializer {
           updatedInputs,
           spendingTransaction.outputs,
           spendingTransaction.lockTime)
-        val sigHashBytes = hashType.num.bytes.reverse
+        val sigHashBytes = Int32(hashType.num).bytes.reverse
 
         hashType match {
           case _: SIGHASH_NONE =>
@@ -213,7 +213,8 @@ sealed abstract class TransactionSignatureSerializer {
           spendingTransaction.version.bytes.reverse ++ outPointHash ++ sequenceHash ++
             i.previousOutput.bytes ++ CompactSizeUInt.calc(scriptBytes).bytes ++
             scriptBytes ++ amount.bytes ++ i.sequence.bytes.reverse ++
-            outputHash ++ spendingTransaction.lockTime.bytes.reverse ++ hashType.num.bytes.reverse
+            outputHash ++ spendingTransaction.lockTime.bytes.reverse ++ Int32(
+              hashType.num).bytes.reverse
         serializationForSig
     }
   }
